version dropdown in sidebar
