+++ components/navbar.js
@@ -11,14 +11,12 @@ import { toast } from 'react-toastify';
 import BridgeSlider from './sliders/bridgeSlider';
 import ChatBotSlider from './sliders/chatBotSlider';
 import OrgSlider from './sliders/orgSlider';
-import { getVersionHistoryAction } from '@/store/action/historyAction';
-import { setSelectedVersion } from '@/store/reducer/historyReducer';
+
 
 function Navbar() {
   const router = useRouter();
   const searchParams = useSearchParams()
   const versionId = searchParams.get('version')
-  const thread_id = searchParams.get('thread_id')
   const dispatch = useDispatch();
   const pathName = usePathname();
   const path = pathName.split('?')[0].split('/')
@@ -30,11 +28,6 @@ function Navbar() {
     bridge: state.bridgeReducer.allBridgesMap[bridgeId] || [],
     publishedVersion: state?.bridgeReducer?.allBridgesMap?.[bridgeId]?.published_version_id || [],
     isdrafted: state?.bridgeReducer?.bridgeVersionMapping?.[bridgeId]?.[versionId]?.is_drafted,
-  }));
-  const params = useParams();
-
-  const { selectedVersion } = useCustomSelector((state) => ({
-    selectedVersion: state?.historyReducer?.selectedVersion || 'all'
   }));
 
   const handleDeleteBridge = async (item, newStatus = 0) => {
@@ -101,27 +94,6 @@ function Navbar() {
   const toggleBridgeSidebar = () => toggleSidebar('default-bridge-sidebar');
   const toggleChatbotSidebar = () => toggleSidebar('default-chatbot-sidebar');
 
-
-
-  const { bridgeVersionsArray } = useCustomSelector(
-    (state) => ({
-      bridgeVersionsArray: state?.bridgeReducer?.allBridgesMap?.[params?.id]?.versions || [],
-    })
-  );
-
-  const handleVersionChange = async (event) => {
-    const version = event.target.value;
-    dispatch(setSelectedVersion(version));
-
-    if (version !== "all") {
-      try {
-        dispatch(getVersionHistoryAction(thread_id, params.id, version))
-      } catch (error) {
-        console.error('Failed to fetch version history:', error);
-      }
-    }
-  };
-
   return (
     <div className={` ${pathName === '/' || pathName.endsWith("alerts") ? 'hidden' : 'flex items-center justify-between '} w-full navbar border flex-wrap md:flex-nowrap z-[100] max-h-[4rem] bg-base-100 sticky top-0 mb-3`}>
       <div className={`flex items-center w-full justify-start gap-2 ${path.length > 4 ? '' : 'hidden'}`}>
@@ -171,28 +143,6 @@ function Navbar() {
                 >
                   Publish Version
                 </button>
-                <div className="divider divider-horizontal mx-1"></div>
-              </div>
-            )}
-            {path[4] === 'history' && (
-              <div className='flex items-center'>
-
-                <select
-                  className="select select-bordered w-full max-w-xs"
-
-                  value={selectedVersion}
-                  onChange={handleVersionChange}
-                >
-                  <option value="all">All Versions</option>
-                  {bridgeVersionsArray?.map((version, index) => (
-                    <option
-                      key={version}
-                      value={version}
-                    >
-                      Version {index + 1}
-                    </option>
-                  ))}
-                </select>
                 <div className="divider divider-horizontal mx-1"></div>
               </div>
             )}
